Cookbook: move generator to solution top level
