+++ src/ScottPlot4/ScottPlot.Cookbook.Generator/Program.cs
@@ -1,0 +1,21 @@
+﻿/* This application generates the images and Markdown pages used to build the cookbook website.
+ * This application is intended to be executed from inside Visual Studio.
+ */
+
+namespace Cookbook_Generator;
+
+public static class Program
+{
+    public static void Main()
+    {
+        string cookbookProjectFolder = Path.GetFullPath("../../../../ScottPlot.Cookbook");
+        string outputFolder = Path.GetFullPath("../../../../ScottPlot.Cookbook/CookbookOutput");
+
+        ScottPlot.Cookbook.RecipeSource[] recipes = ScottPlot.Cookbook.Generator.Generate(
+            cookbookProjectFolder, outputFolder, regenerate: true);
+
+        Website.Generate(outputFolder, recipes);
+
+        Console.WriteLine($"COMPLETE: {outputFolder}");
+    }
+}
