+++ src/ScottPlot4/ScottPlot.Cookbook.Generator/ScottPlot.Cookbook.Generator.csproj
@@ -8,7 +8,7 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<ProjectReference Include="..\..\ScottPlot.Cookbook\ScottPlot.Cookbook.csproj" />
+		<ProjectReference Include="..\ScottPlot.Cookbook\ScottPlot.Cookbook.csproj" />
 	</ItemGroup>
 
 </Project>
