+++ src/ScottPlot4/ScottPlot.Cookbook.Generator/Template.cs
