+++ src/ScottPlot4/ScottPlot.Cookbook.Generator/Website.cs
